requirements.txt
fastapi
uvicorn
openai
python-dotenv
gspread
oauth2client
